Add option to keep Grappolo clusters. And do not require true partition info.

+++ StochasticBlockPartition/code/python/partition_grappolo.py
@@ -13,6 +13,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument("input_filename", nargs="?", type=str, default="../../data/static/simulated_blockmodel_graph_500_nodes")
     parser.add_argument("-t", "--threads", type=int, required=False, default=0)
+    parser.add_argument("-k", "--keep", type=int, required=False, default=0)
     args = parser.parse_args()
 
     path_to_grappolo = '~/grappolo/grappolo-05-2014/driverForGraphClustering'
@@ -22,7 +23,11 @@ if __name__ == '__main__':
         print("File doesn't exist: '{}'!".format(input_filename))
         sys.exit(1)
 
-    true_partition = np.loadtxt(input_filename + '_truePartition.tsv', dtype=np.int64)[:,1] - 1
+    true_partition_available = True
+    try:
+        true_partition = np.loadtxt(input_filename + '_truePartition.tsv', dtype=np.int64)[:,1] - 1
+    except:
+        true_partition_available = False
 
     if args.threads != 0:
         os.environ["OMP_NUM_THREADS"] = str(args.threads)
@@ -32,7 +37,10 @@ if __name__ == '__main__':
     print(out)
 
     cluster_output_fname = input_filename + '.tsv_clustInfo'
-    partition = np.loadtxt(cluster_output_fname, delimiter='\t', dtype=np.int64)
 
-    evaluate_partition(true_partition, partition)
-    os.remove(cluster_output_fname)
+    if true_partition_available:
+        partition = np.loadtxt(cluster_output_fname, delimiter='\t', dtype=np.int64)
+        evaluate_partition(true_partition, partition)
+
+    if not args.keep:
+        os.remove(cluster_output_fname)
